Add docstrings for agents and refactor agent
