+++ erniebot-agent/src/erniebot_agent/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from erniebot_agent.utils.logging import logger, setup_logging
+from erniebot_agent.utils.logging import setup_logging
 
 __all__ = ["logger"]
