+++ erniebot-agent/src/erniebot_agent/agents/functional_agent.py
@@ -14,7 +14,7 @@
 
 from typing import List, Optional, Union
 
-from erniebot_agent.agents.base import Agent, ToolManager
+from erniebot_agent.agents.agent import Agent
 from erniebot_agent.agents.callback.callback_manager import CallbackManager
 from erniebot_agent.agents.callback.handlers.base import CallbackHandler
 from erniebot_agent.agents.schema import AgentAction, AgentFile, AgentResponse
@@ -29,23 +29,60 @@ from erniebot_agent.memory.messages import (
     SystemMessage,
 )
 from erniebot_agent.tools.base import BaseTool
+from erniebot_agent.tools.tool_manager import ToolManager
 
 _MAX_STEPS = 5
 
 
 class FunctionalAgent(Agent):
+    """An agent driven by function calling.
+
+    The orchestration capabilities of a functional agent are powered by the
+    function calling ability of the integrated LLMs. Typically, a functional
+    agent asks the LLM to generate a response that can be parsed into an action
+    (e.g., calling a tool with given arguments), and then the agent takes that
+    action, which forms an agent step. The agent repeats this process until the
+    maximum number of steps is reached or the LLM considers the task finished.
+
+    Attributes:
+        llm: The LLM that the agent uses.
+        memory: The message storage that keeps the chat history.
+        max_steps: The maximum number of steps in each agent run.
+    """
+
     def __init__(
         self,
         llm: ChatModel,
         tools: Union[ToolManager, List[BaseTool]],
         memory: Memory,
-        system_message: Optional[SystemMessage] = None,
         *,
+        system_message: Optional[SystemMessage] = None,
         callbacks: Optional[Union[CallbackManager, List[CallbackHandler]]] = None,
         file_manager: Optional[FileManager] = None,
-        plugins: Optional[List[str]] = None,  # None is not assigned, [] is no plugins.
+        plugins: Optional[List[str]] = None,
         max_steps: Optional[int] = None,
     ) -> None:
+        """Initialize a functional agent.
+
+        Args:
+            llm: An LLM for the agent to use.
+            tools: A list of tools for the agent to use.
+            memory: A memory object that equips the agent to remember chat
+                history.
+            system_message: A message that tells the LLM how to interpret the
+                conversations. If `None`, the system message contained in
+                `memory` will be used.
+            callbacks: A list of callback handlers for the agent to use. If
+                `None`, a default list of callbacks will be used.
+            file_manager: A file manager for the agent to interact with files.
+                If `None`, a global file manager that can be shared among
+                different components will be implicitly created and used.
+            plugins: A list of names of the plugins for the agent to use. If
+                `None`, the agent will use a default list of plugins. Set
+                `plugins` to `[]` to disable the use of plugins.
+            max_steps: The maximum number of steps in each agent run. If `None`,
+                use a default value.
+        """
         super().__init__(
             llm=llm,
             tools=tools,
@@ -82,8 +119,8 @@ class FunctionalAgent(Agent):
             )
             if curr_step_output is None:
                 response = self._create_finished_response(chat_history, actions_taken, files_involved)
-                self.memory.add_message(chat_history[0])
-                self.memory.add_message(chat_history[-1])
+                self._memory.add_message(chat_history[0])
+                self._memory.add_message(chat_history[-1])
                 return response
             num_steps_taken += 1
             next_step_input = curr_step_output
@@ -113,7 +150,7 @@ class FunctionalAgent(Agent):
     ) -> Optional[AgentAction]:
         chat_history.append(input_message)
 
-        messages = self.memory.get_messages() + chat_history
+        messages = self._memory.get_messages() + chat_history
         llm_resp = await self._async_run_llm(
             messages=messages,
             functions=self._tool_manager.get_tool_schemas(),
