+++ erniebot-agent/src/erniebot_agent/agents/schema.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import annotations
-
 import functools
 from dataclasses import dataclass
 from typing import Dict, List, Literal, Optional, Tuple, Union
