+++ erniebot-agent/src/erniebot_agent/tools/utils.py
@@ -1,5 +1,6 @@
 import base64
 import inspect
+import logging
 import typing
 from copy import deepcopy
 from typing import Any, Dict, Optional, Type
@@ -18,7 +19,8 @@ from erniebot_agent.tools.schema import (
 )
 from erniebot_agent.utils.common import get_file_suffix, is_json_response
 from erniebot_agent.utils.exceptions import RemoteToolError
-from erniebot_agent.utils.logging import logger
+
+logger = logging.getLogger(__name__)
 
 
 def tool_response_contains_file(element: Any):
